extract db out
